requirements.txt
asgiref==3.10.0
certifi==2025.11.12
cffi==2.0.0
cryptography==46.0.3
dj-database-url==3.0.1
Django==5.2.8
django-cors-headers==4.9.0
django-debug-toolbar==6.1.0
django-extensions==4.1
django-filter==25.2
django-redis==6.0.0
django-rest-registration==0.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dotenv==0.9.9
drf-yasg==1.21.11
gunicorn==23.0.0
inflection==0.5.1
packaging==25.0
pillow==12.0.0
psycopg2-binary==2.9.11
pycparser==2.23
pygraphviz==1.14
PyJWT==2.10.1
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
redis==7.1.0
sqlparse==0.5.3
typing==3.7.4.3
uritemplate==4.2.0
whitenoise==6.11.0
